cleanup around draft and message search
