+++ src/utils/AsyncStore.js
@@ -1,5 +1,7 @@
 import AsyncStorage from '@react-native-community/async-storage';
 
+import ChatClientStore from './ChatClientStore';
+
 export default {
   flushGetRequests: () => AsyncStorage.flushGetRequests(),
   getAllKeys: () => AsyncStorage.getAllKeys(),
@@ -18,3 +20,14 @@ export default {
     await AsyncStorage.setItem(key, JSON.stringify(value));
   },
 };
+
+const chatClient = ChatClientStore.client;
+
+export const getChannelDraftKey = (channelId) =>
+  `@slack-clone-draft-${chatClient.key}-${chatClient.user.id}-${channelId}`;
+
+export const getUserDraftKey = () =>
+  `@slack-clone-draft-${chatClient.key}-${chatClient.user.id}`;
+
+export const getRecentSearchesKey = () =>
+  `@slack-clone-recent-searches-${chatClient.key}-${chatClient.user.id}`;
