--- src/utils/draftUtils.js
@@ -1,7 +1,0 @@
-import ChatClientStore from './ChatClientStore';
-
-const chatClient = ChatClientStore.client;
-export const getChannelDraftKey = (channelId) =>
-  `@slack-clone-draft-${chatClient.key}-${chatClient.user.id}-${channelId}`;
-export const getUserDraftKey = () =>
-  `@slack-clone-draft-${chatClient.key}-${chatClient.user.id}`;
